chore: move positional embeddings into pos_embed
